requirements.txt
telebot==3.8.1
python-docx==0.8.11
Aspose.Words==1.0.0